The path to the executable might have spaces in it.  Guard against this by quoting the variable.
(Unfortunately, this means that a trick I was using a while ago to add
options to the command line by overriding the LARCENY environment
variable will not work.  To regain that expressive power, I guess I
will later have to change the definitition of the COMPOPTS setting to
make sure that it inherits its value from the environment.  But this
is not a concern for right now, I think...)


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5058 c9681fa6-b8ae-40c0-a6da-707a35ad5973