One possible fix for Ticket #492.
The problem described in the ticket was non-deterministic because it
depended on the result of the {{{.GetHashCode()}}} method in {{{C#}}}.

The essential problem is that Will is using the expression
{{{(remainder h n)}}} in the hashtable code, where {{{h}}} is a
hashcode and {{{n}}} is the number of buckets in the table.  But the
{{{remainder}}} operation will then yield a result that inherits the
sign of its {{{h}}} argument, and thus if the hashcode is negative,
the {{{remainder}}} operation will fail.

There are two possibile fixes that PnkFelix sees for this problem.

One is to change the hashtable code to use {{{modulo}}} instead of
{{{remainder}}}.  However, {{{modulo}}} is not a primop in Larceny
(only {{{remainder}}} is), and so I suspect Will will not want to use
it as a fundamental operation in the hashtable implementation.

The other fix, implemented here, is to ensure that the
{{{machine-address}}} primop always returns a non-negative integer.

Perhaps '''both''' fixes should be implemented; after all, the
hashtable code will be relying on hashcodes that users pass in.  Then
again, the R6RS section on the {{{hashtables}}} library says that
''hash-function'' should return a non-negative exact integer object,
so I guess that puts the responsibility on the client to provide a
hash function that does not return values outside of that range.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5071 c9681fa6-b8ae-40c0-a6da-707a35ad5973