Added a hyperlink at a point where there is a reference to another section of the manual.  I am doing this at one isolated point so that Will can clearly see what I did and try using it himself.
I could have written <<Performance>> for the cross reference, but then
that is rendered as the text "[Performance]" in the resulting text, which
looked ugly to me.  So I went with <<Performance,Performance section>>. 

Perhaps there is a way to actually render the section number, which
would be better for our printed version of the documentaton.

There are a couple different choices for how to write the anchor and
the crossref (aka xref).  I am tempted to stop using the markup syntax
{{{<<}}}''id''{{{,}}}''caption''{{{ >>}}} and 
{{{[[}}}''id''{{{,}}}''xreflabel''{{{]]}}}, because it is not intuitively
obvious what the role of each of the two syntaxes is.  The alternative
forms {{{xref:}}}''id''{{{[}}}''caption''{{{]}}} and 
{{{anchor:}}}''id''{{{[}}}''xreflabel''{{{]}}} are more verbose, but
they immediately say what their respective roles are.

For the official reference on this, see "Internal Cross References" in
this section of the AsciiDoc manual:

http://www.methods.co.nz/asciidoc/asciidoc.css-embedded.html#_inline_macros


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5118 c9681fa6-b8ae-40c0-a6da-707a35ad5973