Fix for problem with dynamic region allocation.
While cheney is chugging away on {{{gclib_stopcopy_collect}}}, 
it might attempt to allocate a generation number, and shift 
all of the other generations around to accomodate it.  So 
the {{{set_gen_no}}} method in {{{old-heap}}} sets the 
generation number of the semispace in the heap's 
{{{current_space}}} field.

But at the time when cheney is invoked, the {{{current_space}}} 
field hadn't been updated with the target {{{tospace}}} yet.

It seems safe to update the {{{current_space}}} field before 
invoking cheney.  (At the very least, it fixes the problem
where the dynamic area was getting a generation number that
overlapped the generation numbers of the ephemeral areas.)
So lets adopt that change and see how things go.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5195 c9681fa6-b8ae-40c0-a6da-707a35ad5973