Changed collection policy to ensure that old regions are always collected before new ones.
I did this because the collector was easily getting into a situation
where it would start with:                   regions { A, B }

where objects point from A into B and thus are keeping elements of B
alive.

while collecting B, create fresh region C:   regions { A, B, C }
while collecting C, create fresh region D:   regions { A, B, C, D }

and so on, so that we never get around to collecting A, and thus
the garbage in B never gets collected.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5292 c9681fa6-b8ae-40c0-a6da-707a35ad5973