While looking over the changesets on the [source:branches/pnkfelix/markthread markthread branch] and comparing against the current code, I realized that something must have gone awry on the trunk because this code is no longer updating the {{{ms_promotion}}} stats fields.
The problem was the refactoring I attempted to perform in
changeset:4890; it should have differentiated between whether it was
stopping timers for a promotion or for a collection, and it did not do
so.  

This changeset fixes that, and should be ported over to trunk when
someone (me) gets the chance.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5293 c9681fa6-b8ae-40c0-a6da-707a35ad5973