Added a remembered summarization pass. (This is analogous to changeset:4366 on the [source:branches/pnkfelix/markthread markthread branch].)
I decided to implement this because it is the easiest path to
supporting the major/minor remset distinction.

After doing it, I realized that I had put the summary construction
code outside the timers for remset-scanning (see the {{{cheney.c}}}
portion of changeset:5294), and therefore the remset scan timing code
is now only capturing the time spent:

1. Looking at all of the remsets for a minor collection
2. Scanning the remset summary itself.

(1.) above is going to dominate until I reintroduce the major/minor
remset distinction.  But I can emulate the effects of that by doing a
remset summarization that *only* looks at the nursery.  

So I think I'm going to spend some time doing that emulation and then
seeing what sort of remset scan times we get then.  If they continue
to be unreasonable in that situation (due to absurdly large
summaries), then I may have to take a step back and rethink which task
is most important for me to attack.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5298 c9681fa6-b8ae-40c0-a6da-707a35ad5973