This changeset emulates the effect of introducing the major/minor remset distinction by performing a prepass construction of a summary for just the nursery; the assumption is that scanning a collection of minor remsets takes time comparable to scanning such a summary.
With this change, running nboyer:3,

with 10 megabyte regions, max remset scan 369ms, avg remset scan  8.7ms
with  5 megabyte regions, max remset scan 291ms, avg remset scan 17.9ms

Instrumentation indicates that the summary sometimes contains as many
as 600,000 objects on the 10M run, and 800,000 objects on the 5M run.
Those are the extreme cases, of course, and it is cases like that
which are leading us towards trying to handle the popular object
problem.  The question is, should I be seeing this kind of behavior on
this benchmark?  Or is this a sign that something else is wrong?


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5300 c9681fa6-b8ae-40c0-a6da-707a35ad5973