Calculate the number of outside-referrers via mark/sweep visitor and print that as well, so that we can see how much it differs from what the remset summary tells us for that value...
The number of outside referrers via mark/sweep could be an
over-estimate, since (I think) for any object O with k refs (to <= k
objects) into the region, the mark visitor will add k to the
popularity analysis, while O should contribute 1 entry to the summary.

But this presents all the more reason that it is very odd that the
summary is growing *larger* than the amount calculated by the
mark visitor (let alone so much larger)!


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5308 c9681fa6-b8ae-40c0-a6da-707a35ad5973