Macro to enqueue a word onto the SSB.
It has only been mildly tested.

(I was trying a quick hack of trying to speed up cheney by having it
use the SSB instead of the remset routines.  The hack did not pay off,
but I may need something similar in the future, so I am adding the
macro.  The right next step might be to try replacing other SSB
enqueuing code with a use of the macro...)


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5309 c9681fa6-b8ae-40c0-a6da-707a35ad5973