Revised popularity analysis code.
It now builds up a remset for the marker, so that one can directly
investigate where "extra entries" in the remset summary are coming
from.

It also categorizes the membership of the remset summary.  Using that,
I have determined that the extra entries in the remset summary
correspond to now dead objects.  (This should not have been a
surprise...)

To better understand this behavior, I added an additional
categorization of the dead members of summary remset, where the
analysis prints out what the breakdown is across the various regions
of how many dead entries each region is contributing.

All of this can (and should) be disabled by removing/disabling
invocations of the {{{popularity_analysis}}} function.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5318 c9681fa6-b8ae-40c0-a6da-707a35ad5973