Bug fix to mistake I made back when I did the initial prototype support for forwarding into multiple semispaces (in [source:branches/pnkfelix/markthread/larceny_src/src/Rts/Sys/cheney.c markthread cheney.c]).
For any semispace ss, the chunks at indices beyond {{{ss->current}}}
are not "allocated" to the garbage collectable heap; they are merely
being held in reserve for future allocation.  Therefore, it is not
valid to scan them as part of the tospace; you are only supposed to
scan up to (and including) {{{ss->current}}}.  

The previous code was scanning up to {{{ss->n}}}, which would almost
always work due to the way that semispaces and the chunks within them
are currently allocated.  But my attempts to experiment with semispace
reuse were being thwarted by this bug, since the collector would
attempt to scan chunks that were not allocated as part of the tospace
(only for reserve in the tospace), and encounter dead bits such as old
forwarding pointers.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5334 c9681fa6-b8ae-40c0-a6da-707a35ad5973