Some beta code that avoids a sequence of semispace operations: {{{[free; create; expand*; free; create; expand*; ...]}}}
This code (when enabled via the {{{ATTEMPT_TO_REUSE_SEMISPACES}}} flag)
instead caches the last freed semispace, and then returns it as 
the result of a compatible invocation of the {{{create_semispace}}}
constructor.

The hardest thing about this change was tracking down the bug it
exposed (which I then fixed in changeset:5334).

It is not turned on by default because I first want to confirm that it
is a win overall, and also confirm that there is not a saner approach
achieving the same effect via local changes to the
[source:branches/pnkfelix/dynregion/larceny_src/src/Rts/Sys/old-heap.c
old-heap.c] or 
[source:branches/pnkfelix/dynregion/larceny_src/src/Rts/Sys/memmgr.c 
memmgr.c] client code.

Also note that this code could lead to the runtime holding onto memory
that should have been returned to the operating system, since a
semispace's chunks won't be freed until another semispace is freed
(without an intervening compatible semispace creation).  I have
thoughts on approaches to this, such as freeing some portion of the
chunks on the initial {{{ss_free}}} invocation, and thus over time the
semispace should approach a steady state, assuming that all semispaces
are periodically passed to {{{ss_free}}}.  But for now that is all
just hypothetical, since this code is still beta quality.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5335 c9681fa6-b8ae-40c0-a6da-707a35ad5973