Some beta code that avoids repeated {{{malloc}}}'ing and {{{free}}}'ing of pool chunks within a remset.
This code (when enabled via the {{{ATTEMPT_TO_REUSE_POOL_SEGMENTS}}}
flag) instead stores the unused pool segments for a remset in another
linked list, and then grabs new segments from that pool instead of
always using {{{allocate_pool_segment}}}.

This is even more beta quality than the semispace changes in
changeset:5335.  In particular, when {{{ATTEMPT_TO_REUSE_POOL_SEGMENTS}}} 
is turned on, the storage associated with a remembered set is '''never'''
returned to the operating system.  
This is almost certainly something that would have to be fixed before
I could turn this flag on for production code.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5336 c9681fa6-b8ae-40c0-a6da-707a35ad5973