A new entry point for the mark/sweep library routines, {{{msgc_mark_objects_from_roots_and_remsets}}}, which treats all of the remembered set(s) as a source of roots.
This is not the appropriate entry point for most uses of mark/sweep,
since usually one is deliberately using mark/sweep to avoid the float
that is being held due to (now dead) remembered set entries.

But some of my recent experiments with using the mark/sweep system as
the basis for an oracle that can artificially maintain heap invariants
(such as updating the remembered sets after cheney has finished)
specifically require that I treat the remembered set as a set of
roots.  Otherwise the oracle will not traverse objects that are not
reachable from the roots but '''are''' reachable from the rememembered
sets and thus '''will''' be traversed during some cheney
copying-collection in the future.

(This code has not been explicitly tested yet.)


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5337 c9681fa6-b8ae-40c0-a6da-707a35ad5973