uses new msgc entry point to ensure soundness of results provided by oracle.  (This is the pessimistic approach; an alternative optimistic approach would be to pro-actively clean dead objects from the remembered sets.)
Note that the oracle is quadratic (and always has been).  This is not
acceptable, even for experimentation.  I can get around the problem
easily for minor collections (by adding a different oracle that works
by simply scanning the tospace).  A similar approach might be
reasonable for major collections (the main issue there is identifying
which semispaces actually belong to the tospace, since the act of
collecting may have allocated new regions; but that should not be too
hard to deal with).


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5340 c9681fa6-b8ae-40c0-a6da-707a35ad5973