There was a serious oversight in the implementation I had for {{{find_space_rgnl}}}; the {{{gc_allocated_to_areas}}} method does not return correct results in the midst of collection.
In particular, it would use the number of allocated bytes at the
''start'' of collection.  Thus it would claim that that region
associated with the first tospace was filled up when in fact we had
only filled up a single chunk of a semispace; it would then create a
fresh space, and then subsequent invocations of {{{find_space_rgnl}}}
would claim that the fresh space has nothing allocated to it, even
though the collector had been forwarding objects into it.  (Again,
this is because it would use the number of allocated bytes at the
''start'' of collection). 

So we would end up with a region gradually growing larger than the
bounds we are trying to impose on them, and thus the pause time would
not be bounded.

This change is an attempt to fix this, by doing the same calculation
that is used to set the {{{allocated}}} field of an {{{old-heap}}}
structure to compute the value of {{{cur_allocated}}} for the region.

This fixes the bad pause times I was seeing on the perm9 benchmark.  I
still need to evaluate how other benchmarks are doing...


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5355 c9681fa6-b8ae-40c0-a6da-707a35ad5973