Some experimental code to use the {{{load_factor}}} to predict the amount of storage needed in the future and preallocate regions for them.
Let '''L''' be an estimate of the current live storage.  (Will and I
agreed to calculate this estimate by the sum of live data in each of
the regions at their last major collection; that is what the
{{{live_last_major_gc}}} field in the {{{old_heap_t}}} structure is
for, since I am using {{{old_heap_t}}} to represent each region.)

Let '''M''' be the total amount of storage accomodated by the current
set of regions.  (This is the sum of the {{{maximum}}} fields for each
of the {{{old_heap_t}}} structures.)

Let '''P''' be the amount of storage we need to accomodate future live
data.  If '''P''' exceeds '''M''', then we create fresh regions (each
of which obviously contributes nothing to '''L''' and increases
'''M''') until '''P''' is less than '''M'''.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5367 c9681fa6-b8ae-40c0-a6da-707a35ad5973