Added code to do a full mark pass (and then refine the remembered sets based on the mark information) every ''k'' collections (where ''k'' is a parameter to the {{{-mark_period}}} command line option).
Also added code to measure the amount of float in each region, and a
flag to control whether we perform this measurement and then print the
results to the console.

Finally, moved the oracular remset verification pass during
{{{before_collection}}} to the end of the procedure, so that the
implicit mark it performs will traverse the flushed stack frames.  
(At the start of the procedure, the young heap's
{{{yh_before_collection}}} method has not yet been invoked, and
therefore the stack cache has not been flushed to the heap.)


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5389 c9681fa6-b8ae-40c0-a6da-707a35ad5973