Revised instrumentation for measuring float in the regions.
It attempts to isolate float that is guaranteed to be collected in the
next major collection of that region (as opposed to float that
requires a regional cycle to complete, or a mark remset-refinement to
complete).  Such float is marked with a {{{Z}}} in the output bars,
and all other float is marked with an {{{R}}}.

However, the measurement of {{{Z}}} is grossly underestimated in the
current changeset, because the instrumentation gathers its estimate of
all other float by doing a mark pass over the heap where '''all''' of
the remembered sets are used as roots, but a major collection of
region ''R'' will ''not'' use the (points-out) remembered set for
''R'' as a set of roots, and thus there are often many objects that
will be included in the major collection that will not be put into the
{{{Z}}} category above.

(I do believe the calculation of {{{Z}}} is at least a conservative
underestimate; any object categorized as {{{Z}}} really will be
collected on the next major collection.  I would just prefer to make
the {{{Z}}} calculation more accurate.)


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5435 c9681fa6-b8ae-40c0-a6da-707a35ad5973