Slight change to how the internal accounting to schedule a mark.
Before, a long series of minor collections could cause a
mark-refinement to be scheduled, but the extra nursery evacuations
that happened ''after'' the scheduling but ''before'' the actual
mark-refinement would not be accounted for.  This meant that we were
marking less often than we should (at least according to the
user-specified refinement policy), since there were several megabytes
of allocation that should have been credited towards a
mark-refinement.

Fixing this by removing the invariant that the countdown be
non-negative.  Now, a negative countdown represents the number of
nursery evacuations that occurred since the mark-refinement was
scheduled.  This is treated as credit towards a future mark
(that's why the {{{rrof_refine_mark_countdown}}} reinitialization
 in {{{refine_remsets_via_marksweep}}} uses a {{{+=}}} rather than just 
 an {{{=}}}: the negative present value is treated as a credit towards
 an earlier marking pass.)


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5440 c9681fa6-b8ae-40c0-a6da-707a35ad5973