Another experimental change.
This one attempts to stop the collector from being fooled by float
into increasing the heap size to absurd levels.  The countermeasure is
to continually average the new live storage estimate (calculated as
the sum of survivors of major collections) with the most recent
estimate of live storage (which is either the prior calcuation or the
words marked in the last mark pass; whichever has happened more
recently).

The weight given to the prior estimate is a preprocessor definition
{{{WEIGH_PREV_ESTIMATE_LOADCALC}}}; note that it is set to {{{0}}}
here, which should make this changeset have no observable effect on
the existing behavior.

  * That is, in this changeset {{{live_estimated_calc}}} should always
    be set to {{{total_live_at_last_major_gc}}} and so the computed
    {{{live_predicted_at_next_gc}}} (sic) should come out the same as
    before unless you change {{{WEIGH_PREV_ESTIMATE_LOADCALC}}} to
    some positive integer.)


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5441 c9681fa6-b8ae-40c0-a6da-707a35ad5973