Fix for a problem that I noticed when I was reviewing the last changeset.
The purpose of subtracting the nursery size is to counter the case of
equilibrium where we're continually promoting the whole nursery into a
region A while a nursery's worth of objects are dying in region
B... in this situation, including the nursery in the heap size
estimate is a mistake and will lead to an over-estimate of the live
storage.

Subtracting the nursery size ensures that we should always calculate
an under-estimate (assuming that the marking process is running often
enough to keep the remembered sets precise enough).

But subtracting the nursery size when we promote less than a nursery's
amount of objects can lead to a region claiming that it contributes a
negative amount to the total heap size, which is bogus and leads to
gross under-estimates of the live storage, which in turn leads to
insufficient heap expansion and then we waste a lot of time copying
float.

I am currently addressing this problem by ensuring that each region
always claims to contribute a non-negative amount to the total heap
storage.  It would probably be better to precisely calculate the
number of promotions out of the nursery and then ''only'' subtracting
that from the live storage in the region.

(The problem described above happened in practice, in particular on
20earley:13 with 5 MB regions and a refinement of 5.0 -- this
changeset reduces its elapsed time from 73sec to 48sec)


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5469 c9681fa6-b8ae-40c0-a6da-707a35ad5973