This assertion was being violated in cases where view was scrolled to the far right for a single long line and then the user deletes sufficient characters from the line that the horizontal offset (represented as {{{first-column-idx}}}) falls beyond the length of the line itself.
The right way to handle this would probably be to "automatically
scroll", e.g. by adjusting the {{{first-column-idx}}} so that it falls
within the new max line length.  (At least, that is how Windows
Notepad behaves.)

But another (seemingly) reasonable solution, that I have adopted here,
is maintain the invariant by ensuring that the max value for the
horizontal scrollbar is always kept above {{{first-column-idx}}}, even
when the max line length drops below that value.  The user has to
explicitly choose to adjust the horizontal scrollbar back (to the far
left) before the scrollbar will ever go away.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5537 c9681fa6-b8ae-40c0-a6da-707a35ad5973