Sketch of button toolbar support for editor.
Seems to work for the single simple test case I have made, so I
figured it was worth checking in before I left for the day.

Here is that simple test case:
{{{(load "lib/Experimental/DotNet/example.sch") ;; as usual
   ((editor-wnd 'push-buttons) 
    (make-btn 'title "Hi" 'make-agent make-noisy-agent)     ;; 1st button
    (make-btn 'title "Hello" 'make-agent make-noisy-agent)) ;; 2nd button
   (start-example) ;; as usual
}}}

The first biggest pain is that I could not find a simpler way to
handle component layout in an automatic fashion.  (Here I am treating
the buttontoolbar as an ever-present entity; the fact that it does not
show up when there are no buttons is an artifact of the fact that it
is given zero height when it has no buttons.)

The second biggest pain was the core control (that is, the text
editor) does not automatically receive focus when the user clicks on
them, so I added an explicit mouse down handling event that invokes
the Control.Focus() method.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5545 c9681fa6-b8ae-40c0-a6da-707a35ad5973