Fixing dependency information in our generated Makefiles.
I built an immediate dependency matrix by grepping for occurrences of
{{{#include}}}.  I then compared that with a matrix generated by
observing the actual {{{make}}} behavior after touching each header
file in turn.  I then combined the two matrices to make a third matrix
that had all of the dependences which, according to our use of
{{{#include}}}, should be reflected in the {{{make}}} behavior, but
are not.  That process gave me a set of dependences to add to the make
file.

The other issue is how to handle dependences that are due to indirect
uses of {{{#include}}} (that is, a header that included another header
file).  Here is the policy I am trying to follow:

If a header file uses {{{#include}}}, then the generated Makefile
should define a make variable for it.  That variable should be set to
header file itself and all of the immediately included files (or
variables for such included files, in the case of included files that
themselves use {{{#include}}}).  The only place where an including
header file should be be referenced in the Makefile is within the
definition of its variable; all other references should be to the make
variable.

For example, {{{barrier.h}}} includes {{{larceny-types.h}}} and
{{{gclib.h}}}.  Therefore define a variable {{{BARRIER_H}}} that is
set to {{{Sys/larceny-types.h}}} and {{{$(GCLIB_H)}}}.

(This follows a precedent set by Lars in his use of a {{{LARCENY_H}}}
variable for a large set of headers that are all {{{#include}}}'d by
{{{larceny.h}}}.)


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5609 c9681fa6-b8ae-40c0-a6da-707a35ad5973