Expanded foreign-cstructs library to be more flexible and to infer more information.
The old version of foreign-cstructs was a very small veneer over what
foreign-ctools provided; for example, it required one to provide a
low-level, size-dependent %getX routine to extract state, and a %setX
routine to set it.  This was not necessary for simple case where all
one wants is integer values, since foreign-ctools now can infer
{{{sizeof}}} information of a structure's field.  Even in the more
general case where one does not want an integer but rather some more
abstract value, the right thing to do there is to encourage the user
to use the FFI attribute functionality and provide a name for the
data they are trying to extract.

(There are cases where this code has actually become too abstract.  In
particular, I believe the old code was able to handle it when strings
were embedded within structures, since its conversion routine received
the offset into the struct bytevector, and could read off the string
from there.  But for right now I think it is better to provide the
high level version for these cases, and let people dealing with
embedded strings use {{{define-c-info}}} directly to get at the
necessary addresses and offsets.)


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5685 c9681fa6-b8ae-40c0-a6da-707a35ad5973