My own attempt to combat assembler overhead (Ticket #575).
This code had inherited some weird artifacts from the NASM backend in
[source:trunk/larceny_src/src/Asm/Nasm/pass5p2-nasm.sch Nasm/pass5p2-nasm.sch].
In particular, it built up a list of seen labels that it actually did not need.
(The list of seen labels was a superset of the labels stored with the
assembly structure itself)

Then I made a mistake and used the seen labels as the basis of a list
of assembly symbols to export from a segment (so that we can get the
machine code offsets for each label after Sassy finishes assembling
it).  We did not need all of those seen labels to be exported; we only
need the labels that are stored with the {{{as-labels}}} field.

So this change eliminates the seen labels slot entirely from the data
structure particular to the Sassy backend, and builds up the list of
symbols to export from just the {{{as-labels}}} field that all of the
assemblers share.

As a happy semi-accident, this improves the space usage of the
assembler by an incredible amount.

Here the max memory data produced by loading {{{boo.sch}}} before this change:
{{{
 Mem:     words 175734784
 Heap:    words 171113472
 Remset:  words 149504
 Rts:     words 17408
 Waste:   words 3950592
}}}

Here's the same statistics produced after this change:
{{{
 Mem:     words 14558208
 Heap:    words 14179328
 Remset:  words 124928
 Rts:     words 17408
 Waste:   words 207872
}}}

That's a factor of 10 reduction in max memory usage!!!


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5722 c9681fa6-b8ae-40c0-a6da-707a35ad5973