Ticket #583 -- mzscheme 37x cannot read {{{#vu8()}}} syntax.
Does the wiki document the different classes of files in the Larceny
source tree?  In between Larceny-specific code (which I believe lives
in {{{lib/}}} and in {{{src/Compat/Larceny}}} and semi-portable Scheme
code, there is also Scheme code which must be portably readable but
can contain invocations of Larceny-specific procedures.  (There may be
additional classes based on what special forms can be used.)


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5737 c9681fa6-b8ae-40c0-a6da-707a35ad5973