Small revision to the marking variant that includes the remsets as roots.  (This variant is only used for internal assertions and debugging, not for the actual collection algorithms.)
Obviously roots have no source object, so I used to pass {{{0x0}}} as
the source parameter.  But I have found that it is sometimes useful to
know for which remembered set we are pushing an object onto the mark
stack, so I am now propagating that information via a static variable.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5743 c9681fa6-b8ae-40c0-a6da-707a35ad5973