Added instrumentation to track when new logging sequential store buffers (SSB) are allocated.
When an SSB is allocated, the associated word log structure may also
need to be reallocated, and the other SSBs may need to be relinked
with bot/top/lim for their new logs.  So I am adding
{{{seqbuf_swap_in_ssb}}} to allow such dynamic relinking of word logs.

The sp_data associated with an SSB's callback may also need to change
in response to changes elsewhere in the runtime, so I am adding that
function as well.  The particular case this handles for the RROF
collector is: each SSB log is associated with a region, and when the
SSB is flushed it needs to both add its entries to its region's
summary (if present) and distribute its entries amongst the remembered
sets (always).  The association between a log and its region is
represented by the region's gno, but a region's gno may ''change''
when new regions are constructed.  So when that change happens, 
it has to propogate to the SSB as well.

Anyway, I'll be shortly adding descriptions for both of these
functions in the seqbuf_t.h header file.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5752 c9681fa6-b8ae-40c0-a6da-707a35ad5973