Switching cheney so that it uses the callback method provided by the {{{gc_t}}} object instead of the previously hard-coded enqueue of the lhs word into its remembered set.
The remset enqueue code was ported to the callback method in
changeset:5747 so this is ''almost'' a pure refactoring.

Here's why it is not a pure refactoring: the previous code would
return TRUE from {{{points_across}}} when enqueuing {{{lhs}}} (which
the cheney code would use as a signal that it could switch to a
tighter object scanning loop; see the {{{scan_update_rs}}} macro in
[source:branches/dynregion/larceny_src/src/Rts/Sys/cheney.h cheney.h]
of the dynregion branch, looking in particular at the invocations of
{{{UPDATE_REMSET}}}).

This speed-hack is not sound in the RROF algorithm, because the object
scanning loop may need to enqueue the object X into the summaries of
other regions ''as well as'' the remembered set of X's region, and
therefore we cannot switch to the faster loop just because X happened
to be enqueued into a remembered set.  E.g., if objects A, B, and C
live in regions 1, 2, and 3, and A{{{[1]}}} points to B and A{{{[2]}}}
to C, then the scan of A{{{[1]}}} may first enqueue A in the
remembered set for region 1 and also into the summary for region 2,
but the scanning loop still may need to enqueue A into the summary for
region 3 when it sees A{{{[2]}}}.  Therefore we cannot switch to the
tighter scanning loop immediately after scanning A{{{[1]}}}.

What all of this adds up to is: I need to review the effects of this
change on the speed characteristics of the code, comparing the GC
performance of the incrsumz branch with the GC performance of the
dynregion branch.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5754 c9681fa6-b8ae-40c0-a6da-707a35ad5973