Fixing a bug in internal (debug) instrumentation.
When analyzing the amount of data and float in a region, I was
accidentally rounding the number of ''words'' occupied by 
{ vector, bytevector, procedure } objects up to a multiple of 8. 

The standard pattern we use in e.g. {{{cheney.c}}} is to first round
the number of ''bytes'' up to a multiple of 8, and then divide that by
the word size in bytes.  You can see from the diff how getting your
order of operations right can make a big difference (especially if
lots of small objects are running around).


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5781 c9681fa6-b8ae-40c0-a6da-707a35ad5973