Fix for an old (and subtle) bug in remembered set refinement.
The problem is essentially that I was not refining the remembered set
for the static area after doing a mark of the objects reachable from
the roots.

The reason that this is a problem is that you end up with objects in
the static area that are in the static area's remembered set, but are
themselves unreachable and have references to other unreachable
objects in non-static areas of the heap.  The other objects are then
collected, which makes the references in the static dead objects
invalid.  But since the static area's remembered set was not being
refined, we would still attempt to '''scan''' the unreachable objects
in the static area, which would lead to attempts to dereference
dangling pointers (bad).


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5789 c9681fa6-b8ae-40c0-a6da-707a35ad5973