Refactored code to remove the {{{rgn_idx}}} variable, replacing it with {{{rgn_next}}} and {{{rgn_to}}} as appropriate.
The intention is that {{{rgn_to}}} is always the ''to space'' (during
both major and minor collections) and {{{rgn_next}}} is the ''from
space'' region during a major collection.

In the current code base, a major collection is fired when
{{{rgn_to}}} == {{{rgn_next}}}, which is why I was able to get away
with collapsing them both into {{{rgn_idx}}}.  But I worked two weeks
ago to allow major collections to fire when {{{rgn_to}}} !=
{{{rgn_next}}}, and a big part of that was teasing out the different
uses of {{{rgn_idx}}} into the two original variables in this code.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5799 c9681fa6-b8ae-40c0-a6da-707a35ad5973