Somewhat "simplified" version of my code that decouples TO region (collection target) from NEXT region (major collection source) during major collections.
The desired policy for choosing when to do minor and when to do major
is still pretty up in the air; the code currently in here is pretty
much a crock.

I just wanted to get this commited to SVN before some disaster befalls
my laptop.  (I also want to move on to doing work on implementing real
SATB-style mark-and-refinement, since I suspect that the information
gleaned from the SATB marking passes could be crucial components in
guiding the heap expansion policy, which will affect much any policy
for a decoupled next/to.)


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5802 c9681fa6-b8ae-40c0-a6da-707a35ad5973