Switching internal representation for the mark stack so that it uses a single mark stack rather than a stack-per-region.
This may seem like a step backward, but at this point I think it will
be more difficult to make the ''stack-per-region'' approach
competitive with a single mark stack than it will be to update a
single stack in response to collector activity.  In particular, each
region has a linked list threaded through the stack entries, so that
we only need to update the individual entries related to a region,
rather than scanning the entire stack.

(The reason that ''stack-per-region'' was posing difficulties is that
I was no longer able to reason easily about how large the mark stacks
were growing in the expected case.)


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5815 c9681fa6-b8ae-40c0-a6da-707a35ad5973