A note about something I want to investigate later.
Over the weekend I noticed some problems where the estimated heap size
was much lower than it should be, and therefore the heap expansion
policy was not expanding the heap as much as one would like.  Maybe
all this will go away with an expansion policy based (mostly) around
the SMIRCY results.

But the point is that right now, the value of the
{{{live_last_major_gc}}} field may not be trust-worthy.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5822 c9681fa6-b8ae-40c0-a6da-707a35ad5973