Removed unnecessary code to update incremental marker state.
I am pretty sure that the removed code was an artifact of an earlier
non-incremental version of the mark-and-refine code, where the roots
where pushed to start things off.  The current code should not need
these lines because the roots should be taken at the time the snapshot
is created, and the write-barrier should handle adding any updates
that the marker needs to know about.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5908 c9681fa6-b8ae-40c0-a6da-707a35ad5973